added insitu subpakage and moved code - currently broken
