+++ msbrainpy/insitu/_gene_series.py
@@ -1,0 +1,7 @@
+
+# File contains functions specific to processing images in a series in situ
+# brain series to resolve information about the brains in the series and the 
+# gene in question.
+# -----------------------------------------------------------------------------
+# Signal quantification
+# ---------------------
